23092020-1833 From DEV to MASTER

+++ integridad-api/src/main/java/com/mrzolution/integridad/app/domain/DetailOffline.java
@@ -21,7 +21,7 @@ public class DetailOffline implements Child {
     @GeneratedValue
     private UUID id;
     
-    private long quantity;
+    private Double quantity;
     private Double costEach;
     private Double total;
     private String adicional;
